refactor(k8s-provider): extract AutoScalingWorkload as a sibling shape
The StatefulSet logic lived as private methods bolted onto the 800-line
KubernetesProvider next to the static Deployment machinery. Lift it into a
cohesive AutoScalingWorkload unit so the two shapes are siblings, not a patch
on one — the facade (apply/start/stop/destroy/observe/observeAll) becomes a
thin dispatcher: apply branches on runtimeMode; the spec-less lifecycle/read
methods detect the shape from what exists.

The EnvironmentProvider interface is unchanged — the runtime shape is spec
data (runtimeMode), not a new operation, so reconcile-core stays shape-
agnostic and a future non-k8s provider is a new impl, not a core change.

- support.ts: shared helpers (resourceName, workspaceLabels, createOrAdopt,
  swallow404, expandWorkspacePvc) used by both shapes.
- auto-scaling-workload.ts: AutoScalingWorkload over injected api clients +
  config, self-contained and fake-testable.
- provider.ts: dispatch to this.autoScaling; static Deployment path unchanged.

No behavior change: tsc clean, 220 cp unit tests pass, golden-master pod
templates byte-identical.
